+++ src/main/java/it/unibz/inf/ontouml/vp/model/ServerRequest.java
@@ -1,4 +1,4 @@
-package it.unibz.inf.ontouml.vp.utils;
+package it.unibz.inf.ontouml.vp.model;
 
 public class ServerRequest implements Runnable {
